tests/requirements.txt
Django~=1.11
djangorestframework~=3.9
pytest-django~=3.5
